Finished user inventory item modal, polished admin item modal
