+++ rss-client/src/app/Cart/select-cart/select-cart.component.scss
@@ -78,7 +78,7 @@ a:hover {
 
 .img-small {
     width: 40px;
-    height: 40px;
+    object-fit: cover;
 }
 
 .img-frame {
