+++ rss-client/src/app/Cart/shopping-cart/shopping-cart.component.scss
@@ -26,9 +26,7 @@
 }
 
 @media screen and (min-width: 770px) {
-    .center-divide {
-        border-right: 2px solid #474C55;
-    }
+
 }
 
 @media screen and (max-width: 770px) {
