+++ rss-client/src/app/app-inventory/components/add-item/add-item.component.scss
@@ -1,13 +1,8 @@
-.btn-outline-orange{
-    color: #F26925 !important;
-    border-color: #F26925 !important;
-}
-.btn-outline-orange:hover{
-    color: #fff !important;
-    background-color: #F26925 !important;
-    border-color: #F26925 !important;
-}
 .text-sm{
     font-size: 12px;
     margin-top: 3px;
+}
+
+label {
+    font-weight: bold;
 }
