+++ rss-client/src/app/app-inventory/components/inventory-item/inventory-item.component.scss
@@ -1,0 +1,13 @@
+#product-img {
+    width: 250px;
+    object-fit: contain;
+}
+
+.modal-footer {
+    display: block !important;
+}
+
+label {
+    font-weight: bold;
+    margin-bottom: 3px !important;
+}
