+++ rss-client/src/app/app-inventory/components/user-inventory-item/user-inventory-item.component.html
@@ -6,51 +6,39 @@
 </div>
 
 <div class="modal-body">
-    <div *ngIf="userType==='admin'">
-        <p class="text-sm text-danger mb-1">* These fields are required.</p>
-    </div>
-    <form (ngSubmit)="updateItem()" [formGroup]="updateProduct">
+    <form [formGroup]="updateProduct">
         <div class="row">
             <div class="col-12 col-md-6 mb-2">
-                <span class="text-danger" *ngIf="userType==='admin'">* </span>
-                <strong *ngIf="userType==='admin'; else nameBlock">Name:</strong>
-                <ng-template #nameBlock><span>Name:</span></ng-template>
-                <input type="text" class="form-control form-control-sm" id="inputName" name="name" formControlName="name" required>
-                <p class="text-danger text-sm" [hidden]="name.valid || name.pristine"><strong>Name</strong> is required</p>
+                <b>Name:</b> {{product.name}}
             </div>
-            <div class="col-12 col-md-6 mb-2">Brand:
-                <input type="text" class="form-control form-control-sm" id="brand" name="brand" formControlName="brand">
+            <div class="col-12 col-md-6 mb-2">
+                <b>Category:</b> {{product.category}}
             </div>
-        </div>
-        <div class="mb-2">Description:
-            <input type="text" class="form-control form-control-sm" id="inputDescription" name="description" formControlName="description">
         </div>
         <div class="row">
-            <div class="col-12 col-md-6 mb-2">Model:
-                <input type="text" class="form-control form-control-sm" id="inputModel" name="model" formControlName="model">
+            <div class="col-12 col-md-4 mb-2">
+                <b>Brand:</b> {{product.brand}}
             </div>
-            <div class="col-12 col-md-6 mb-2">Category:
-                <input type="text" class="form-control form-control-sm" id="category" name="category" formControlName="category">
+            <div class="col-12 col-md-4 mb-2">
+                <b>Model:</b> {{product.model}}
+            </div>
+            <div class="col-12 col-md-4 mb-2">
+                <b>Color:</b> {{product.color}}
             </div>
         </div>
         <div class="mb-2">
-            <img [src]="image.value" alt="product_img" class="rounded mx-auto d-block mt-4 mb-4" style="width: 250px;">
-            <div *ngIf="userType==='admin'">
-                Image URL:
-                <input type="text" class="form-control form-control-sm" id="inputImage" name="image" formControlName="image">
-            </div>
+            <b>Description:</b> {{product.description}}
+        </div>
+        <div class="mb-2">
+            <img [src]="image.value" id="product-img" alt="product_img" class="rounded mx-auto d-block mt-4 mb-4">
         </div>
         <div class="row">
-            <div class="col-12 col-md-6 mb-2">
-                <span class="text-danger" *ngIf="userType==='admin'">* </span>
-                <strong *ngIf="userType==='admin'; else quantityBlock">Stock Quantity:</strong>
-                <ng-template #quantityBlock><span>Stock Quantity:</span></ng-template>
+            <div class="col-12 col-md-4 mb-2">
+                <b>Quantity:</b>
                 <input type="number" class="form-control form-control-sm" id="quantity" name="quantity" formControlName="quantity" required min="0" pattern="([0-9]+)">
-                <!-- <input type="number" class="form-control form-control-sm" required min="0" pattern="([0-9]+)"> -->
                 <p class="text-danger text-sm" [hidden]="quantity.valid || quantity.pristine">
                     <strong>Quantity</strong> is required
                 </p>
-
                 <div class="alert alert-danger alert-dismissible fade show" role="alert" *ngIf="negativeQ">
                     <strong>Sorry!</strong> Try adding fewer items to your cart.
                     <button type="button" class="close" data-dismiss="alert" aria-label="Close">
@@ -58,43 +46,20 @@
 					</button>
                 </div>
             </div>
-            <div class="col-12 col-md-6 mb-2">
-                <span class="text-danger" *ngIf="userType==='admin'">* </span>
-                <strong *ngIf="userType==='admin'; else pointsBlock">Points:</strong>
-                <ng-template #pointsBlock><span>Points:</span></ng-template>
-                <input type="number" class="form-control form-control-sm" id="unitPrice" name="unitPrice" formControlName="unitPrice" required min="0" pattern="([0-9]+)">
-                <p class="text-danger text-sm" [hidden]="unitPrice.valid || unitPrice.pristine">
-                    <strong>Points</strong> is required
-                </p>
-            </div>
-            <!--
-				For this Branch we have adjusted calculations in the shopping cart to make discount inputs not a subtraction operation,
-				but instead a percentage based calculation.
-            -->
-            <!-- admin block for editting discount -->
-            <div class="col-12 col-md-6 mb-2" *ngIf="userType==='admin'">
-                <strong>Percent Off:</strong>
-                <input type="number" class="form-control form-control-sm" id="discountedAmount" name="discountedAmount" formControlName="discountedAmount" pattern="([0-9]+)">
-            </div>
-
-            <!-- customer block for viewing discounted product price -->
-            <div class="col-12 col-md-6 mb-2" *ngIf="userType ==='customer' && product.discounted">
-                <strong>Percent Off:</strong>
-                <input type="number" class="form-control form-control-sm" id="discountedAmount" name="discountedAmount" formControlName="discountedAmount" pattern="([0-9]+)">
+            <div class="col-12 col-md-4 mb-2">
+                <b>Points: </b><span [style.text-decoration]="product.discounted ? 'line-through' : 'none'">{{product.unitPrice}}</span>
+                <span *ngIf="product.discounted === true" style="color:#FF0000">{{product.discountedAmount | number}}% OFF!!!!! {{product.unitPrice - (product.unitPrice * (product.discountedAmount*0.01))}}</span>
             </div>
         </div>
-        <div class="mb-2">Color:
-            <input type="text" class="form-control form-control-sm" id="inputColor" name="color" formControlName="color">
-        </div>
-
-        <div *ngIf="userType==='admin'" class="modal-footer">
-            <button type="button" class="btn btn-danger" (click)="activeModal.dismiss('Cross click')">Cancel</button>
-            <button type="submit" class="btn btn-primary" [disabled]="!updateProduct.valid">Save</button>
-        </div>
-        <div *ngIf="userType==='customer'" class="modal-footer">
-            <button type="button" class="btn btn-danger" (click)="activeModal.dismiss('Cross click')">Cancel</button>
-            <button type="button" class="btn btn-primary" (click)="reduceInventory()">Add to Cart</button>
+        <div class="modal-footer">
+            <div class="row">
+                <div class="col-6">
+                    <button type="button" class="btn btn-danger btn-block" (click)="activeModal.dismiss('Cross click')">Cancel</button>
+                </div>
+                <div class="col-6">
+                    <button type="submit" class="btn btn-primary btn-block" (click)="reduceInventory()">Add to Cart</button>
+                </div>
+            </div>
         </div>
     </form>
-    {{updateProduct.value | json}}
 </div>
