+++ rss-client/src/app/app-inventory/components/user-inventory-item/user-inventory-item.component.scss
@@ -1,0 +1,8 @@
+#product-img {
+    width: 250px;
+    object-fit: contain;
+}
+
+.modal-footer {
+    display: block !important;
+}
